bevrand.ocrapi/requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pytesseract==0.2.9
pytest==5.0.1
pytest-cov==2.6.1
Werkzeug==0.15.5
